jintan/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-stubs==1.7.0
djangorestframework==3.12.2
djangorestframework-stubs==1.4.0
flake8==3.8.4
idna==2.10
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.800
mypy-extensions==0.4.3
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
